train_neuro/context/requirements.txt
#wheel==0.34.2
#numpy==1.24.2
#--extra-index-url https://download.pytorch.org/whl/cu113
#torch 
#torchvision
#torchaudio
numpy
scikit-learn
Pillow
matplotlib
seaborn
tqdm
pydot

